+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - `Formal\AccessLayer\Row::toArray()`
+- Allow to specify the connection charset via the `charset` query parameter in the connection `Url`